+++ src/main/java/com/example/seoultechInvestment/DTO/EnrollDTO.java
@@ -7,6 +7,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -17,9 +18,9 @@ public class EnrollDTO {
     @NotBlank(message = "종목명이 입력되지 않았습니다.")
     private String tickerName;
     @NotNull(message = "tp값이 입력되지 않았습니다.")
-    private double tp;
+    private BigDecimal tp;
     @NotNull
-    private double entryPrice; // 추천 진입가
+    private BigDecimal entryPrice; // 추천 진입가
 
     private LocalDate enrollDate;
 }
